+++ core/data/src/main/java/com/dmdiaz/currency/core/data/datasources/RatesLocalDataSource.kt
@@ -27,7 +27,7 @@ package com.dmdiaz.currency.core.data.datasources
 
 import arrow.core.Either
 import com.dmdiaz.currency.core.database.entities.DBRates
-import com.dmdiaz.currency.libs.models.Failure
+import com.dmdiaz.currency.core.domain.models.Failure
 import kotlinx.coroutines.flow.Flow
 
 import org.joda.money.CurrencyUnit
