--- core/data/src/main/java/com/dmdiaz/currency/core/data/di/DispatchersModule.kt
@@ -1,50 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2024 David Diaz
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package com.dmdiaz.currency.core.data.di
-
-
-
-import com.dmdiaz.currency.core.data.di.qualifiers.Dispatcher
-import com.dmdiaz.currency.core.data.di.qualifiers.Dispatchers.Default
-import com.dmdiaz.currency.core.data.di.qualifiers.Dispatchers.IO
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.Dispatchers
-
-@Module
-@InstallIn(SingletonComponent::class)
-object DispatchersModule {
-
-    @Provides
-    @Dispatcher(Default)
-    fun provideDefaultDispatcher() : CoroutineDispatcher = Dispatchers.Default
-
-    @Provides
-    @Dispatcher(IO)
-    fun provideNetworkDispatcher() : CoroutineDispatcher = Dispatchers.IO
-}
