+++ core/data/src/main/java/com/dmdiaz/currency/core/data/repositories/RatesRepositoryImpl.kt
@@ -29,10 +29,10 @@ import com.dmdiaz.currency.core.data.datasources.RatesLocalDataSource
 import com.dmdiaz.currency.core.data.mappers.toDBRates
 import com.dmdiaz.currency.core.data.mappers.toRates
 import com.dmdiaz.currency.core.database.converters.DateConverter
-import com.dmdiaz.currency.libs.extensions.mapRight
+import com.dmdiaz.currency.libs.util.extensions.mapRight
 import com.dmdiaz.currency.core.data.datasources.RatesRemoteDataSource
-import com.dmdiaz.currency.core.data.di.qualifiers.Dispatcher
-import com.dmdiaz.currency.core.data.di.qualifiers.Dispatchers.Default
+import com.dmdiaz.currency.libs.util.di.qualifiers.Dispatcher
+import com.dmdiaz.currency.libs.util.di.qualifiers.Dispatchers.Default
 import com.dmdiaz.currency.core.domain.repositories.RatesRepository
 import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.*
