+++ core/data/src/main/java/com/dmdiaz/currency/core/data/repositories/Repository.kt
@@ -26,13 +26,11 @@ package com.dmdiaz.currency.core.data.repositories
 
 import arrow.core.Either
 import arrow.core.raise.either
-import arrow.core.flatten
 import arrow.core.left
 import arrow.core.raise.Raise
-import com.dmdiaz.currency.libs.extensions.filterNotNullRight
-import com.dmdiaz.currency.libs.extensions.tryOrFailure
-import com.dmdiaz.currency.libs.models.Failure
-import com.dmdiaz.currency.libs.models.Failure.UnknownError
+import com.dmdiaz.currency.libs.util.extensions.filterNotNullRight
+import com.dmdiaz.currency.core.domain.models.Failure
+import com.dmdiaz.currency.core.domain.models.Failure.UnknownError
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.FlowCollector
 import kotlinx.coroutines.flow.cancellable
@@ -40,8 +38,7 @@ import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.conflate
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.onEach
-import kotlinx.coroutines.withContext
+import kotlinx.coroutines.flow.map
 
 @Suppress("unused", "MemberVisibilityCanBePrivate", "RedundantAsync")
 abstract class Repository(
@@ -55,37 +52,21 @@ abstract class Repository(
         fetch: suspend Raise<Failure>.() -> FETCH,
         saveFetchSuccess: suspend Raise<Failure>.(FETCH) -> Unit,
     ) = flow(domainFlow)
-        .catch { emit(UnknownError(it).left()) }
-        .onEach { domain ->
+        .map { domain ->
             either{
-                tryOrFailure {
-                    val domainSuccess = domain.bind()
-                    val shouldFetchSuccess =  either { shouldFetch.invoke(this, domainSuccess) }.bind()
-                    if (shouldFetchSuccess){
-                        val fetchSuccess = either { fetch.invoke(this) }.bind()
-                        either { saveFetchSuccess.invoke(this, fetchSuccess)}.bind()
-                    }
-                    domainSuccess
+                val domainSuccess = domain.bind()
+                val shouldFetchSuccess =  either { shouldFetch.invoke(this, domainSuccess) }.bind()
+                if (shouldFetchSuccess){
+                    val fetchSuccess = either { fetch.invoke(this) }.bind()
+                    either { saveFetchSuccess.invoke(this, fetchSuccess)}.bind()
                 }
-            }.flatten()
+                domainSuccess
+            }
         }
+        .catch { emit(UnknownError(it).left()) }
         .filterNotNullRight()
         .flowOn(defaultDispatcher)
         .cancellable()
         .conflate()
-
-    suspend fun <RESPONSE, DOMAIN> crud(
-        operation: Raise<Failure>.() -> RESPONSE,
-        saveOperationSuccess: Raise<Failure>.(RESPONSE) -> DOMAIN,
-    ) = withContext(defaultDispatcher) {
-        either<Failure, DOMAIN> {
-            val success = tryOrFailure { either(operation) }.flatten().bind()
-            tryOrFailure {
-                either<Failure, DOMAIN> {
-                    saveOperationSuccess(success)
-                }
-            }.flatten().bind()
-        }
-    }
 
 }
