+++ core/domain/build.gradle.kts
@@ -27,7 +27,7 @@ plugins {
     alias(libs.plugins.android.library)
     alias(libs.plugins.kotlin.android)
     alias(libs.plugins.hilt.android)
-    kotlin("kapt")
+    alias(libs.plugins.devtools.ksp)
 }
 
 android {
@@ -61,11 +61,10 @@ android {
 
 dependencies {
 
-    implementation(project(":libs:common"))
-
+    implementation(project(":libs:util"))
     //hilt
     implementation(libs.hilt.android)
-    kapt(libs.hilt.android.compiler)
+    ksp(libs.hilt.android.compiler)
 
     //joda
     implementation(libs.joda.money)
