+++ libs/designsystem/build.gradle.kts
@@ -26,12 +26,10 @@
 plugins {
     alias(libs.plugins.android.library)
     alias(libs.plugins.kotlin.android)
-    alias(libs.plugins.hilt.android)
-    alias(libs.plugins.devtools.ksp)
 }
 
 android {
-    namespace = "com.dmdiaz.currency.features.rates"
+    namespace = "com.dmdiaz.currency.libs.designsystem"
     compileSdk = 34
 
     defaultConfig {
@@ -65,13 +63,8 @@ android {
 dependencies {
 
     implementation(project(":libs:util"))
-    implementation(project(":libs:designsystem"))
-
-    implementation(project(":core:domain"))
-
-    //hilt
-    implementation(libs.hilt.android)
-    ksp(libs.hilt.android.compiler)
+    implementation(libs.navigation.fragment.ktx)
+    implementation(libs.navigation.ui.ktx)
 
     implementation(libs.preference.ktx)
     implementation(libs.datastore.preferences)
@@ -83,13 +76,9 @@ dependencies {
     implementation(libs.arrow.core)
     implementation(libs.arrow.fx.coroutines)
 
-
     implementation(libs.core.ktx)
     implementation(libs.appcompat)
     implementation(libs.material)
-    implementation(libs.constraintlayout)
-    implementation(libs.navigation.fragment.ktx)
-    implementation(libs.navigation.ui.ktx)
     testImplementation(libs.junit)
     androidTestImplementation(libs.ext.junit)
     androidTestImplementation(libs.espresso.core)
