+++ libs/designsystem/src/main/AndroidManifest.xml
