+++ libs/designsystem/src/main/java/com/dmdiaz/currency/libs/designsystem/adapters/BaseAdapter.kt
