+++ libs/util/build.gradle.kts
@@ -31,7 +31,7 @@ plugins {
 }
 
 android {
-    namespace = "com.dmdiaz.currency.features.rates"
+    namespace = "com.dmdiaz.currency.libs.util"
     compileSdk = 34
 
     defaultConfig {
@@ -57,39 +57,24 @@ android {
     kotlinOptions {
         jvmTarget = "1.8"
     }
-    buildFeatures {
-        viewBinding = true
-    }
 }
 
 dependencies {
 
-    implementation(project(":libs:util"))
-    implementation(project(":libs:designsystem"))
-
-    implementation(project(":core:domain"))
+    //joda
+    implementation(libs.joda.money)
 
     //hilt
     implementation(libs.hilt.android)
     ksp(libs.hilt.android.compiler)
 
-    implementation(libs.preference.ktx)
-    implementation(libs.datastore.preferences)
-
-    //joda
-    implementation(libs.joda.money)
-
     //arrow
     implementation(libs.arrow.core)
     implementation(libs.arrow.fx.coroutines)
 
-
     implementation(libs.core.ktx)
     implementation(libs.appcompat)
     implementation(libs.material)
-    implementation(libs.constraintlayout)
-    implementation(libs.navigation.fragment.ktx)
-    implementation(libs.navigation.ui.ktx)
     testImplementation(libs.junit)
     androidTestImplementation(libs.ext.junit)
     androidTestImplementation(libs.espresso.core)
