+++ libs/util/src/main/AndroidManifest.xml
