+++ libs/util/src/main/java/com/dmdiaz/currency/libs/util/di/qualifiers/Dispatcher.kt
@@ -22,17 +22,15 @@
  * SOFTWARE.
  */
 
-package com.dmdiaz.currency.core.domain.repositories
+package com.dmdiaz.currency.libs.util.di.qualifiers
 
-import arrow.core.Either
-import com.dmdiaz.currency.core.domain.models.Failure
-import com.dmdiaz.currency.core.domain.models.rates.Rates
-import kotlinx.coroutines.flow.Flow
-import org.joda.money.CurrencyUnit
-import java.util.Date
+import javax.inject.Qualifier
 
-interface RatesRepository {
+@Qualifier
+@Retention(AnnotationRetention.RUNTIME)
+annotation class Dispatcher(val dispatcher: Dispatchers)
 
-    fun getRates(date: Date, currencyUnit: CurrencyUnit): Flow<Either<Failure, Rates>>
-
+enum class Dispatchers {
+    Default,
+    IO,
 }
